+++ app/controllers/integral/backend/posts_controller.rb
@@ -27,7 +27,7 @@ module Integral
       private
 
       def resource_params
-        permitted_post_params = %i[title slug body description tag_list image_id preview_image_id status lock_version user_id]
+        permitted_post_params = %i[title slug body description tag_list image_id preview_image_id status lock_version user_id category_id]
 
         permitted_post_params.concat Integral.additional_post_params
         params.require(:post).permit(*permitted_post_params)