+++ db/migrate/20190929191412_add_integral_post_categories.rb
@@ -1,0 +1,13 @@
+class AddIntegralPostCategories < ActiveRecord::Migration[5.1]
+  def change
+    create_table :integral_categories do |t|
+      t.string :title
+      t.string :description
+      t.string :slug, unique: true
+
+      t.timestamps null: false
+    end
+
+    add_reference :integral_posts, :category, index: true
+  end
+end